Add compressed/head_64.o before vmImage
I tried to jump to 64-bit directly under 1MB, but I failed.
Because the 6 page tables are too big to fit under 1MB's segment.

So I have to add the compressed/ directory to concatenate the head_64.o
with the real vmImage.

The compressed/head_64.o is loaded at 0x100000, it will do the transition
to 64-bit mode, and jump to the our real-kernel.

All this ugly phase is due to the fit for grub+linux.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>