[Stack backtrace is working, thanks to stack unwinder]
Print backtrace is actually very simple: you go through the stack
and find the possible kernel addresses. With the FRAME_POINTER option,
the unwinder can tell which is in the real call trace and which are just
stale values in the stack (with a prefix '?').

Without FRAME_POINTER, a lot of information will be missing. So despite
push %rbp and pop %rbp brings more instructions, the much more friendly
debug info with FRAME_POINTER on can make our debugging life easier.