Test APIC and IO-APIC
In the way to make these 2 guy work, I read a lot of the APIC spec
and IO-APIC spec. Got a thorough understanding of them, and now actually
the local APIC is enabled, I just need to load LVT-Timer to make local
timer interrupt work.

The IO-APIC is not setup yet. But I found the function to do so.
Good luck!

Signed-off-by: Yizhou Shan <shan13@purdue.edu>