BUG: Remove schedule() form timer-handler path
<1>
We are not allowed or able to call schedule() inside timer-interrupt.
Because the timer-handler is NOT allowed to enable interrupt at exit,
however the scheduler() will enable interrupt at last.

The scheduler_tick() function should just manipulate preempt_count or
TIF_NEED_SCHED flag.

(The bug shows up during early boot while calibrating timer, by io-apic
and apic code.)

<2>
The actual schedule() will be invoked at the exit point of the interrupt,
which is the retint_kernel in entry_64.S

This is not finished.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>