Add CONFIG_TRAMPOLINE_START, make the address fixed
We used to use the setup image's base address, which is passed
by bootparam. This will have a problem is the reported address
is not aligned to PAGE_SIZE.

If it is not aligned to PAGE_SIZE, then the smpboot code will
fail to boot the secondary cpus because of the realmode segment
thing I think.

Anyway, to make life easier and easy-to-debug, just make it fixed.
Default to 0x90000 now.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>