Patch initial_gs for per-cpu variable in head.S; Add preempt to spinlock
the initial_gs of BSP is really an interesting.
It is here to make the per-cpu related variable work before setup per-cpu.

Basically it is a trick of linker script, which will produce another symbol
as the real virual address of the per-cpu variable as if it is not per-cpu..

Well

Signed-off-by: Yizhou Shan <shan13@purdue.edu>