Remove tmep_gdt_page, use INIT_PER_CPU(cpu_gdt_page) instead
Once again, a game about linker script.
Variables that are defined as PER_CPU, but have to be used before
per-cpu is setup, have to use this INIT_PER_CPU() trick.

We changed initial_gs in last commit, we changed cpu_gdt_page in this
commit. We only need these two right now.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>