sched_setaffinity() for migrating tasks
FAT NOTE:

A kernel thread is allowed to migrate if and only if:
	1) It is queued on the rq
	2) It is NOT running on a cpu now

So a thread, e.g. A, can NOT call sched_setaffinity() directly itself now.
There must another thread, e.g. B, calls sched_setaffinity(pid_of_A, mask)
for A.

All due to that we are not able to kick-off a thread while it is running
on a cpu. Linux has some migration helper threads that will do this job.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>