Bug fix: Make sched_setaffinity() work in non-preemption kernel
If preemption is NOT configured, then after we move the
task from the old cpu rq to new new cpu, we need to let the new
cpu do a schedule(), so that the new task can get a chance to run.

In most cases, the new cpu is running cpu_idle(). So we need to set
the TIF_NEED_RESCHED in new cpu's idle thread.

We set this after the migrating.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>