[CFS-Patch] Add CFS scheduler to Lego
It is the same CFS scheduler as the one in Linux. Currently, we are able to
change the default scheduler used in [init/init_task.c]. But we are NOT able to
change the scheduler policy at runtime, since we don't have the API yet.

This CFS is tested, and it has several tuning knobs, which are in task_faic.c
All threads have NICE=0, so time is evenly distributed or shard by different threads.