Fix handle_llc_miss(), find_vma is not enough
Note that find_vma() returns the first vma that addr < vm_end.
BUT it does not guarentee addr > vm_start.

This is okay for stack, because stack growsdown, it is okay
to have address < vm_start during runtime.

The semantic we need here is more like find_extended_vma().
So fix the handler now.

Signed-off-by: Yizhou Shan <shan13@purdue.edu>