Clear PF_KTHREAD during exec() call
We need to clear PF_KTHREAD flag for any threads that calls exec().
exec() means run a user program, PF_KTHREAD means a kernel thread.
These 2 babies are opposite.

This bug shows up today when I ran Phoneix word_count. During runtime,
it calls sys_clone() to create a new thread. But the new thread fails
immediatly in the middle of ret_from_fork.

And I saw the failed RIP is actually the newsp passwd in sys_clone().
So copy_process_tls() MUST falely think this is a kernel thread, due
to PF_KTHREAD.

So here we are, fix this bug in sys_exec().

Actually, sys_exec() also need to deal with signal, we leave it as TODO

Signed-off-by: Yizhou Shan <shan13@purdue.edu>