pcache/evict: check PcacheValid before trylock_pcache
If the pcache is usable or within the lru list, it just means this
pcache has been allocated. However, it maybe used by another core's
pgfault code to establish pte mapping (common_do_fill_page), within
which rmap ops will be invoked to lock the pcache.

So, we can not take pcache if it is not Valid yet. Both random and lru
are updated.