[PATCH 1/2] pcache/exit: exited thread need to wait pcache activies
When a thread exit, exit_mm() will set tsk->mm=NULL. Meanwhile, there
are some pcache flush, sweep threads going to use this tsk->mm.

exit_mm() need to wait until all pcache activies done.

Hmm. when a thread exit, how can we remove it from the rmap list???
It is kind of costly to walk all of them. But conceptually, we need to do so.