pcache/rmap: check rmap empty, and better sweep lru
We should not do the rmap walk if rmap list is empty, otherwise we may
end up with a inifinite looping.

Also, in sweep lru function, we should check if pcache is mapped, we
should not try to do the rmap_walk if that is the case, to avoid few
instruction overhead.