memory: Use kzalloc instead of kmalloc in __bprm_mm_init (loader)
This was an potentionl bug that was not triggered previously.
It is simply because kmalloc'ed vma contains some garbage area,
while later in the pgfault code, we use
	if (vma->vm_ops && vma->vm_ops->fault)
		...
to check if it is an file-backed fault.

Fortunately the vma->vm_ops happens to have some leftover value.
So this bug was triggered.

This actually reminds me that this is a serious of potential bugs!
Even though before I've added things like force GFP_ZERO in all
physical page allocation, I missed the kmalloc's case!