Patched unhash_process(), avoid use-after-free bug
This bug is found when testing MT phoenix. I'm so lucky to be able to
find it.

So the root cause of this bug is *laziness*: when you create or allocate
something, you will remember to link or init everything related. Because
if you don't do so, it simply can not run. However, when that something
is being freed, you sometimes forget to cleanup the leftovers. Without
cleaning, others will think this guy still alive and keep using (read or
write) it. But, sine the data structure is freed, it maybe allocated by
kmalloc or buddy again. That means, you will end up corrupting data
silently.

Truly, to be a good system programmer, you need to remember cleanup
whatever you allocated. By cleanup, I mean free the data structure,
update its friends.