pcache: sync refcount between evict and normal users
This is really a nasty fix. But it can cover most race conditions.
The problem can be described as: two threads are using the same pcm, while
one of them is trying to evict it and another is using it.

The evict one is pcache_evict_line. The others can be munmap, mremap, wp
handler. So how we syn between these? We actually need help from two spinlocks
(pte lock and pcache lock), and pcache refcount.

pte lock actually ensure other parties (munmap, mremap, wp) can see a safe
pcm. And the rule is: once drop the pte lock and acquire it again, it must
check if the pte has been changed. If can be unmapped by eviction at the same
time.

Backgroud eviction again live usage is really hard to do. Unlike the Jave GC
which counts refcount to an object, which will not reclaim live objects.
But here, we have the danger of reclaiming live (used by other threads concurrently) pcm.