Processor: User pgtable use per-PTE, per-PMD lock
Before this, all pgtable opearations are protected by one spinlock in mm.
This is bad for multi-threaded applications. We now use per PTE page, and
per PMD page lock.

The spinlock is embedded within `struct page`. The spinlock is 4 bytes.
And as long as `struct page` is not larger than 64 bytes, we are fine.

This optimization applies to Processor only, since it is the one who
manipulate the user pgtables.

Memory probably need some similar stuff. Later.