Reorganize all communication OPCODE and structures
Previously, all opcode and structure are in comp_common.h
This file is used by P, M, and S. This one file solution brings
a lot inconvenience. Some opcode/struct has to be defined twice

To reduce the coding overhead and reduce the probability of writing
mismatched code, we now re-organize things into follows:

- include/lego/rpc/opcode.h: opcode used by P, M, and S
- include/lego/rpc/struct_common: things used by P, M, and S
- include/lego/rpc/struct_p2s: things used by P <-> S communication
- .. others similar

All of them are written in a way that can be included by both Lego
kernel and linux kernel. So in the storage linux module code, it can just
include these files.

For future monitor, please do the same thing.