Processor: add dummy socket/connect
If SOCKET_SYSCALL is not configured in Processor, we create a dummy socket file
operations. Because TensorFlow will try to create a socket, and try to connect,
If TF failed to connect, it will close the socket.

So, if socket does not present, it does not hurt. We just need to provide a dummy
implementation.

But remove this once socket is all tested, and built-in by default.