net: make MAX_OUT be 1 if Processor
If MAX_OUT is > 1, processor is very easy to fail at ib_mad_handler.
Some NULL dereference.

But if we revert it to 1, P will be stable as before.

But I'm still very worried about this handler. It looks like it is
corrupting memory behind out back.

Even though it will not crash if MAX_OUT 1, it does not mean the
handler is corrct in this case.