ExCache Flush: victim will only flush dirty lines
This patch only optimized the victim flush path. Since this should be
the only senario that we will do (even in a real deployment).

The flow is:
When a pcache line chosen to be evicted, it will be unmapped. And check
if any of the rmap PTE is dirty. When the line finished the second step
of insertion to victim, it will be submitted into flush queue.

Right before submit, the dirty flag will be checked. If it is dirty,
the victim line will be flushed. Otherwise, just mark it as Flushed.

nr_victim_flush_submitted_clean
- tell us how many lines have been evicted from pcache in a clean state

nr_victim_flush_submitted_dirty
- tell us how many lines have been evicted from pcache in a drity state