best-effort replication: secondary memory adding flush thread
Implment the best-effort replication design. We have a log flush thread who
is normally sleeping. When secondary memory receives a memory replication
request, it will try to append to log entry to the in-memory log.

If the log is already full or under flush, the current log entry will
be discarded. By doing so, we reduce the impact on foreground performance.

This is the best-effort replication. We just hope the `just discard` page
is no longer useful during recovery.