vicitm: fix double free when there are multiple hits
The bug happen when where are multiple threads hit the same victim
cahche line. The bug can happen because we released the
usable_victim_lock after we do victim_check_hit_entry().

So a faster cpu can drop the fill counter to 0 and think there is no
more users. While another cpu will grab the counter again, also think it
is the only thread using. This eventually leads to double free.

This is the root cause of all recent victim bugs.

This is a confident fix.