rpc_profile: fix the stuck issue in a non-preemptive kernel
Oh well, when is the last time we had issue with non-preemptive kernel.
Oh man, bad memory! The group rule is: with a non-preemptive kernel,
kernel threads will never be "scheduled" away. Kernel threads have to
yield! Otherwise, stuck.

This rpc code, the new created thread happen to run on the same core
as the rpc_profile() thread. And the rpc_profile() is using a busy poll
(while (atomic_read())) to wait for worker.

Anyway, add schedule() to this busy poll waiting.