fit: fix potential bug at wq_handler. reformat code
The wq_handler check list_empty outside of spin_lock.
And this is not correct if we use list_add_tail().
Use list_empty() is only okay if we use list_add().
But this is not recommanded. Always use lock.

Anyway, i add two apis for enqueue and dequeue. And a counter.